Changes in CI scripts

+++ create-local-docker.sh
@@ -13,16 +13,32 @@ stamp=`date +"%s"`
 
 BRANCH_NAME=$(git rev-parse --abbrev-ref HEAD) envsubst <Dockerfile.local > Dockerfile.$stamp
 docker build --rm --network host --build-arg http_proxy=$http_proxy --build-arg https_proxy=$https_proxy -f ${WRKDIR}/Dockerfile.$stamp -t safescale:local-$(git rev-parse --abbrev-ref HEAD | sed 's#/#\-#g') ${WRKDIR}
-[ $? -ne 0 ] && echo "Docker build failed !!" && rm -f ./marker && rm -f ./Dockerfile.$stamp && exit 1
+if [ $? -ne 0 ]; then
+  echo "Docker build failed !!"
+  rm -f ./marker
+  rm -f ./Dockerfile.$stamp
+  exit 1
+fi
 
 echo "Docker build OK"
 
 docker create -ti --name dummy safescale:local-$(git rev-parse --abbrev-ref HEAD | sed 's#/#\-#g') bash
-[ $? -ne 0 ] && echo "Failure extracting binaries 1/3" && exit 1
+if [ $? -ne 0 ]; then
+  echo "Failure extracting binaries 1/3"
+  exit 1
+fi
+
 docker cp dummy:/exported .
-[ $? -ne 0 ] && echo "Failure extracting binaries 2/3" && exit 1
+if [ $? -ne 0 ]; then
+  echo "Failure extracting binaries 2/3"
+  exit 1
+fi
+
 docker rm -f dummy
-[ $? -ne 0 ] && echo "Failure extracting binaries 3/3" && exit 1
+if [ $? -ne 0 ]; then
+  echo "Failure extracting binaries 3/3"
+  exit 1
+fi
 
 rm -f ./Dockerfile.$stamp
